Simplify Compose table implementation
